Code: Update Products request Rules, and changed migration product price from float to decimal

+++ database/migrations/2024_03_06_200724_product.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('description');
-            $table->float('price');
+            $table->decimal('price', 15, 2);
             $table->boolean('status')->default(true);
             $table->timestamps();
             $table->unsignedBigInteger('category_id');
